# specs/004-self-hosted-auth/contracts/api.openapi.yaml
openapi: 3.0.3
info:
  title: Blastoise Authentication API
  description: |
    Self-hosted JWT authentication system replacing Supabase Auth.

    **Security**: All protected endpoints require Bearer token authentication via `Authorization` header.

    **Rate Limiting**:
    - Login: 3 requests/60 seconds per IP
    - Registration: 5 requests/60 seconds per IP
    - Password Reset: 3 requests/60 seconds per email
    - Token Refresh: 10 requests/60 seconds per user

    **Token Lifecycle**:
    - Access tokens: 15 minute TTL
    - Refresh tokens: 7 day TTL
    - Refresh tokens stored in httpOnly cookies
  version: 1.0.0
  contact:
    name: Blastoise API Support
    email: support@blastoise.app

servers:
  - url: http://localhost:3000/api/v1
    description: Local development
  - url: https://api.blastoise.app/v1
    description: Production

tags:
  - name: Authentication
    description: User authentication and session management
  - name: Password
    description: Password reset and management

paths:
  /auth/register:
    post:
      tags:
        - Authentication
      summary: Register new user account
      description: |
        Creates a new user account with email and password. Password is hashed with bcrypt before storage.
        Returns JWT access token and sets refresh token in httpOnly cookie for immediate authentication.
      operationId: register
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
            examples:
              validRegistration:
                summary: Valid registration
                value:
                  email: user@example.com
                  password: SecurePass123
      responses:
        '201':
          description: Registration successful
          headers:
            Set-Cookie:
              description: Refresh token (httpOnly, secure, sameSite)
              schema:
                type: string
                example: refreshToken=eyJhbGc...; HttpOnly; Secure; SameSite=Strict; Max-Age=604800
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
              examples:
                successfulRegistration:
                  summary: Successful registration
                  value:
                    access_token: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
                    token_type: Bearer
                    expires_in: 900
                    user:
                      id: 550e8400-e29b-41d4-a716-446655440000
                      email: user@example.com
                      created_at: '2025-01-18T10:30:00Z'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                weakPassword:
                  summary: Password doesn't meet requirements
                  value:
                    error: Bad Request
                    message: Password must be at least 8 characters and contain at least one letter and one number
                    statusCode: 400
                invalidEmail:
                  summary: Invalid email format
                  value:
                    error: Bad Request
                    message: Invalid email format
                    statusCode: 400
        '409':
          description: Email already registered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                emailExists:
                  summary: Email already in use
                  value:
                    error: Conflict
                    message: This email is already registered
                    statusCode: 409
        '429':
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                rateLimited:
                  summary: Too many registration attempts
                  value:
                    error: Too Many Requests
                    message: Rate limit exceeded. Please try again later.
                    statusCode: 429

  /auth/login:
    post:
      tags:
        - Authentication
      summary: Sign in with email and password
      description: |
        Authenticates user with email and password. Returns JWT access token and sets refresh token in httpOnly cookie.
        Password is verified against bcrypt hash stored in database.
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
            examples:
              validLogin:
                summary: Valid login credentials
                value:
                  email: user@example.com
                  password: SecurePass123
      responses:
        '200':
          description: Login successful
          headers:
            Set-Cookie:
              description: Refresh token (httpOnly, secure, sameSite)
              schema:
                type: string
                example: refreshToken=eyJhbGc...; HttpOnly; Secure; SameSite=Strict; Max-Age=604800
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
              examples:
                successfulLogin:
                  summary: Successful authentication
                  value:
                    access_token: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
                    token_type: Bearer
                    expires_in: 900
                    user:
                      id: 550e8400-e29b-41d4-a716-446655440000
                      email: user@example.com
                      created_at: '2025-01-18T10:30:00Z'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                invalidCredentials:
                  summary: Wrong email or password
                  value:
                    error: Unauthorized
                    message: Invalid email or password
                    statusCode: 401
        '429':
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                rateLimited:
                  summary: Too many login attempts
                  value:
                    error: Too Many Requests
                    message: Too many login attempts. Please try again in 60 seconds.
                    statusCode: 429

  /auth/refresh:
    post:
      tags:
        - Authentication
      summary: Refresh access token
      description: |
        Obtains new access token using refresh token from httpOnly cookie.
        Refresh token is validated and must not be expired or revoked.
        Does NOT rotate refresh token (reusable until revoked).
      operationId: refreshToken
      parameters:
        - name: Cookie
          in: header
          required: true
          description: Refresh token from httpOnly cookie (automatically included by browser)
          schema:
            type: string
            example: refreshToken=eyJhbGc...
      responses:
        '200':
          description: Token refresh successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RefreshResponse'
              examples:
                successfulRefresh:
                  summary: New access token issued
                  value:
                    access_token: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
                    token_type: Bearer
                    expires_in: 900
        '401':
          description: Invalid or expired refresh token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                expiredToken:
                  summary: Refresh token expired
                  value:
                    error: Unauthorized
                    message: Refresh token expired. Please log in again.
                    statusCode: 401
                revokedToken:
                  summary: Token was revoked
                  value:
                    error: Unauthorized
                    message: Invalid refresh token
                    statusCode: 401

  /auth/logout:
    post:
      tags:
        - Authentication
      summary: Sign out and revoke refresh token
      description: |
        Revokes the refresh token associated with the current session.
        Clears refresh token cookie. Access token remains valid until expiration (stateless).
      operationId: logout
      security:
        - BearerAuth: []
      parameters:
        - name: Cookie
          in: header
          required: true
          description: Refresh token to revoke
          schema:
            type: string
            example: refreshToken=eyJhbGc...
      responses:
        '200':
          description: Logout successful
          headers:
            Set-Cookie:
              description: Clears refresh token cookie
              schema:
                type: string
                example: refreshToken=; HttpOnly; Secure; SameSite=Strict; Max-Age=0
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
              examples:
                successfulLogout:
                  summary: User signed out
                  value:
                    message: Successfully logged out
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                notAuthenticated:
                  summary: No valid authentication
                  value:
                    error: Unauthorized
                    message: Authentication required
                    statusCode: 401

  /auth/forgot-password:
    post:
      tags:
        - Password
      summary: Request password reset
      description: |
        Generates a password reset token and sends email with reset link.
        Token is valid for 1 hour and single-use only.
        Returns success even if email doesn't exist (prevents user enumeration).
      operationId: forgotPassword
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ForgotPasswordRequest'
            examples:
              validRequest:
                summary: Valid email
                value:
                  email: user@example.com
      responses:
        '200':
          description: Password reset email sent (or would be sent if email exists)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
              examples:
                emailSent:
                  summary: Reset email sent
                  value:
                    message: If an account exists with this email, you will receive a password reset link.
        '400':
          description: Invalid email format
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                invalidEmail:
                  summary: Malformed email
                  value:
                    error: Bad Request
                    message: Invalid email format
                    statusCode: 400
        '429':
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                rateLimited:
                  summary: Too many reset requests
                  value:
                    error: Too Many Requests
                    message: Too many password reset requests. Please try again later.
                    statusCode: 429

  /auth/reset-password:
    post:
      tags:
        - Password
      summary: Reset password with token
      description: |
        Validates password reset token and updates user's password.
        Token must be valid, not expired, and not already used.
        New password is hashed with bcrypt before storage.
      operationId: resetPassword
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResetPasswordRequest'
            examples:
              validReset:
                summary: Valid reset with new password
                value:
                  token: a1b2c3d4e5f6...
                  new_password: NewSecurePass456
      responses:
        '200':
          description: Password reset successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
              examples:
                passwordReset:
                  summary: Password changed
                  value:
                    message: Password successfully reset. You can now log in with your new password.
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                weakPassword:
                  summary: Password doesn't meet requirements
                  value:
                    error: Bad Request
                    message: Password must be at least 8 characters and contain at least one letter and one number
                    statusCode: 400
        '401':
          description: Invalid or expired token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                invalidToken:
                  summary: Token is invalid or expired
                  value:
                    error: Unauthorized
                    message: Invalid or expired reset token. Please request a new password reset.
                    statusCode: 401
                tokenAlreadyUsed:
                  summary: Token was already used
                  value:
                    error: Unauthorized
                    message: This reset token has already been used
                    statusCode: 401

components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token obtained from login or registration

  schemas:
    RegisterRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          description: User's email address (case-insensitive)
          example: user@example.com
          maxLength: 255
        password:
          type: string
          format: password
          description: User's password (min 8 chars, at least one letter and one number)
          example: SecurePass123
          minLength: 8
          maxLength: 128

    LoginRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          description: User's email address
          example: user@example.com
          maxLength: 255
        password:
          type: string
          format: password
          description: User's password
          example: SecurePass123

    ForgotPasswordRequest:
      type: object
      required:
        - email
      properties:
        email:
          type: string
          format: email
          description: Email address to send password reset link
          example: user@example.com
          maxLength: 255

    ResetPasswordRequest:
      type: object
      required:
        - token
        - new_password
      properties:
        token:
          type: string
          description: Password reset token from email link
          example: a1b2c3d4e5f6g7h8i9j0k1l2m3n4o5p6
          minLength: 32
        new_password:
          type: string
          format: password
          description: New password (min 8 chars, at least one letter and one number)
          example: NewSecurePass456
          minLength: 8
          maxLength: 128

    AuthResponse:
      type: object
      required:
        - access_token
        - token_type
        - expires_in
        - user
      properties:
        access_token:
          type: string
          description: JWT access token (15 minute TTL)
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VyX2lkIjoiNTUwZTg0MDAtZTI5Yi00MWQ0LWE3MTYtNDQ2NjU1NDQwMDAwIiwiZW1haWwiOiJ1c2VyQGV4YW1wbGUuY29tIiwiaWF0IjoxNjQyNTEyMDAwLCJleHAiOjE2NDI1MTI5MDB9.abc123...
        token_type:
          type: string
          enum: [Bearer]
          description: Token type (always "Bearer")
          example: Bearer
        expires_in:
          type: integer
          description: Access token expiration in seconds
          example: 900
        user:
          $ref: '#/components/schemas/UserInfo'

    RefreshResponse:
      type: object
      required:
        - access_token
        - token_type
        - expires_in
      properties:
        access_token:
          type: string
          description: New JWT access token (15 minute TTL)
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        token_type:
          type: string
          enum: [Bearer]
          description: Token type (always "Bearer")
          example: Bearer
        expires_in:
          type: integer
          description: Access token expiration in seconds
          example: 900

    UserInfo:
      type: object
      required:
        - id
        - email
        - created_at
      properties:
        id:
          type: string
          format: uuid
          description: User's unique identifier
          example: 550e8400-e29b-41d4-a716-446655440000
        email:
          type: string
          format: email
          description: User's email address
          example: user@example.com
        created_at:
          type: string
          format: date-time
          description: Account creation timestamp
          example: '2025-01-18T10:30:00Z'

    MessageResponse:
      type: object
      required:
        - message
      properties:
        message:
          type: string
          description: Human-readable message
          example: Operation completed successfully

    ErrorResponse:
      type: object
      required:
        - error
        - message
        - statusCode
      properties:
        error:
          type: string
          description: Error type (HTTP status text)
          example: Bad Request
        message:
          type: string
          description: Human-readable error description
          example: Validation failed
        statusCode:
          type: integer
          description: HTTP status code
          example: 400

security:
  - BearerAuth: []